+++ R/nr-blit.R
@@ -3,10 +3,9 @@
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #' Copy one \code{nativeRaster} image into another at an arbitrary location.
 #'
-#' This is useful as a way of positioning sprites or icons in an image.
+#' Single source blitted to one or more locations.
 #'
-#' @param nr native raster to copy into
-#' @param src native raster to copy from
+#' @param dst,src source and destination native rasters
 #' @param x,y Where in \code{nr} to place the \code{sprite}. These values must
 #'        be vectors of the same length.  If the length is greater than 1, then
 #'        the \code{sprite} will be pasted into \code{nr} at multiple locations.
@@ -17,105 +16,119 @@
 #' @param w,h size within src. If size is negative, then the actual width/height of
 #'        the src is used
 #' @param respect_alpha Should the alpha channel be respected when blitting?
-#'        Default: TRUE.  If FALSE, then contents will be blindly overwritten
-#'        which can be much much faster.  If the \code{src} has an 
-#'        any transparent pixels, \code{respect_alpha = TRUE} is 
-#'        probably the correct setting.
+#'        Default: TRUE means to carefully blend pixels at each location 
+#'        using alpha values.  If FALSE, then contents at the \code{dst} will 
+#'        just be replaced with \code{src} pixels
+#'        which can be much much faster.
 #' @param hjust,vjust specify horizontal and vertical justification of the 
 #'        \code{src} image.  e.g. \code{hjust = vjust = 0} the blitting
 #'        starts at the top-left of the image. Use \code{hjust = vjust = 0.5}
 #'        to treat the centre of the \code{src_} as the blitting origin.
 #'        Default (0, 0)
 #'
-#' @return \code{nativeRaster}
+#' @return None. \code{dst} modifief by-reference and returned invisibly.
 #' 
 #' @examples
 #' nr <- nr_new(50, 50, 'grey80')
-#' nr_blit(nr, x = 0, y = 0, src = deer_sprites[[1]])
+#' nr_blit(dst = nr, src = deer_sprites[[1]], x = 0, y = 0)
 #' plot(nr)
 #' 
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-nr_blit <- function(nr, x, y, src, x0 = 0L, y0 = 0L, w = -1L, h = -1L, hjust = 0, vjust = 0, respect_alpha = TRUE) {
-  invisible(.Call(blit_, nr, x, y, src, x0, y0, w, h, hjust, vjust, respect_alpha))
+nr_blit <- function(dst, src, 
+                    x, y, 
+                    x0    =  0L , y0    =  0L, 
+                    w     = -1L , h     = -1L, 
+                    hjust =  0  , vjust =  0, 
+                    respect_alpha = TRUE) {
+  
+  
+  invisible(.Call(nr_blit_, 
+                  dst  , src, 
+                  x    , y, 
+                  x0   , y0, 
+                  w    , h, 
+                  hjust, vjust, 
+                  respect_alpha))
 }
 
 
+
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Blit from a list of native rasters
+#' Multiple blit operations in a single call
 #' 
-#' @inheritParams nr_blit
-#' @param src_list list of native rasters
-#' @param src_idx indices into the list of the native raster
-#' 
+#' @param dst destination native raster
+#' @param src list of native rasters
+#' @param config data.frame of configuration information for each blit which 
+#'    most contain: idx, x, y, x0, y0, w, h, hjust, vjust, respect_alpha, draw
+#' @return None. \code{dst} modifief by-reference and returned invisibly.
 #' @examples
-#' nr <- nr_new(50, 50, 'grey80')
-#' nr_blit_list(nr, x = c(0, 25), y = c(0, 25), src_list = deer_sprites, src_idx = c(1, 2))
+#' nr <- nr_new(90, 90, 'grey60')
+#' config <- data.frame(
+#'   idx = c(1, 2, 3, 4),
+#'   x = c(10, 10, 40, 40),
+#'   y = c(10, 40, 40, 10),
+#'   x0 = 0L,
+#'   y0 = 0L,
+#'   w = -1L,
+#'   h = -1L,
+#'   hjust = 0,
+#'   vjust = 0,
+#'   respect_alpha = TRUE,
+#'   render = TRUE
+#' )
+#' nr_blit_bulk(dst = nr, src = deer_sprites, config = config)
 #' plot(nr)
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-nr_blit_list <- function(nr, x, y, src_list, src_idx, hjust = 0, vjust = 0, respect_alpha = TRUE) {
-  invisible(.Call(blit_list_, nr, x, y, src_list, src_idx, hjust, vjust, respect_alpha))
-}
-
-
-
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Blit4
-#' 
-#' @inheritParams nr_blit_list
-#' @param idx_mat integer matrix of indices into \code{src_list}
-#' @param width,height tile width/height (constant across all tiles)
-#' 
-#' @return Original \code{nativeRaster} modified in-place
-#' @examples
-#' nr <- nr_new(100, 100, 'grey80')
-#' idx_mat <- matrix(c(
-#'   1, 2, 3,
-#'   4, 5, 6,
-#'   7, 8, 9
-#' ), 3, 3, byrow = TRUE)
-#' nr_blit_grid(nr, 0, 0, src_list = deer_sprites, idx_mat = idx_mat, width = 32, height = 32)
-#' 
-#' @export 
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-nr_blit_grid <- function(nr, x, y, src_list, idx_mat, width, height, hjust = 0, vjust = 0, respect_alpha = TRUE) {
-  
-  if (length(x) != 1 || length(y) != 1) {
-    stop("only allow length 1 for x and y")
-  }
-  
-  if (max(idx_mat, na.rm = TRUE) > length(src_list)) {
-    stop("loc idx is out of bounds")
-  }
-  
-  stopifnot(is.matrix(idx_mat))
-  
-  xoff <- (rep(seq(ncol(idx_mat)), each = nrow(idx_mat)) - 1) * width
-  yoff <- (rep(seq(nrow(idx_mat)),        ncol(idx_mat)) - 1) * height
-  
+nr_blit_bulk <- function(dst, src, config) {
   invisible(
-    nr_blit_list(nr, x = x + xoff, y = y + yoff, src_list = src_list, src_idx = as.vector(idx_mat), 
-             hjust = hjust, vjust = vjust, respect_alpha = respect_alpha)
+    .Call(nr_blit_bulk_, dst, src, config)
   )
 }
 
 
 
 
-
 if (FALSE) {
   
-  logo <- png::readPNG(system.file("img", "Rlogo.png", package="png"), native = TRUE)
-  
-  nr <- nr_new(100, 100, 'grey80')
-  idx_mat <- matrix(c(
-    1, 2, 3,
-    4, 5, 6,
-    7, 8, 9
-  ), 3, 3, byrow = TRUE)
-  nr_blit_grid(nr, 0, 0, src_list = deer_sprites, idx_mat = idx_mat, width = 32, height = 32)
+  logo <- fastpng::read_png(system.file("img", "Rlogo.png", package="png"), type = 'nativeraster')
+  nr <- nr_new(200, 200, 'grey80')
+  nr_blit(nr, logo, 0, 0, hjust = 0.5, vjust = 0.5, respect_alpha = F)
   plot(nr, T)
+  
+  
+  
+  library(grid)
+  
+  # Setup a fast graphics device that can render quickly
+  x11(type = 'cairo', antialias = 'none')
+  dev.control('inhibit')
+  
+  logo <- fastpng::read_png(system.file("img", "Rlogo.png", package="png"), type = 'nativeraster')
+  logo <- nr_scale(logo, 2, algo = 'bilinear')
+  grid <- expand.grid(x = seq(-500, 1000, 220), y = seq(-500, 1000, 200))
+  nr <- nr_new(500, 500, 'white')
+  
+  gc(full = TRUE)
+  dev.flush()
+  dev.flush()
+  dev.flush()
+  gov <- governor::gov_init(1/30)
+  for (i in 1:500) {
+    xoff <- 200 * sin(i/10 * 1/4 * pi)
+    yoff <- 130 * sin(i/13 * 1/4 * pi + pi / 3)
+    nr_fill(nr, 'white')
+    nr_blit(nr, logo, grid$x - xoff , grid$y - yoff, hjust = 0.5, vjust = 0.5, respect_alpha = T)
+    dev.hold()
+    plot(nr)
+    dev.flush()
+    governor::gov_wait(gov)
+  }
+}
+
+
+if (FALSE) {
   
 }