+++ R/nr-core.R
@@ -122,7 +122,10 @@ nr_duplicate <- function(nr) {
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 nr_crop <- function(nr, x, y, w, h) {
   dst <- nr_new(w, h)
-  nr_blit(dst, 0, 0, nr, x, y, w, h, respect_alpha = FALSE)
+  nr_blit(dst, nr, 
+          0, 0, 
+          x, y, 
+          w, h, respect_alpha = FALSE)
   dst
 }