+++ src/nr-blit.c
@@ -19,9 +19,22 @@
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Blit sprite onto raster [C interface]
+// 
+// Function tries to be smart
+//  - trim coordinates to ensure they lie within the src/dst images
+//
+// @param dst destination native raster
+// @param dst_width,dst_height dimensions of destination
+// @param x,y position within destination image
+// @param src source native raster
+// @param src_width,src_height dimensions of source
+// @param x0,y0 starting position within source
+// @param w,h dimensions of region to copy
+// @param respect_alpha Should alpha values be respected when blitting?
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-void blit_core_(uint32_t *dst, int x, int y, int dst_width, int dst_height, 
-                uint32_t *src, int x0, int y0, int w, int h, int src_width, int src_height,
+void blit_core_(uint32_t *dst, int dst_width, int dst_height, int x , int y , 
+                uint32_t *src, int src_width, int src_height, int x0, int y0, 
+                int w, int h,
                 bool respect_alpha) {
   
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -93,9 +106,13 @@ void blit_core_(uint32_t *dst, int x, int y, int dst_width, int dst_height,
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Blit sprite onto raster [C interface]
-// Slow and study, but it works
+// Slow and study, but it works.
+// It attempts to plot every single point regardless of whether it
+// falls inside the boundary or not
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-void blit_core_naive_(uint32_t *dst, int x, int y, int dst_width, int dst_height, uint32_t *src, int x0, int y0, int w, int h, int src_width, int src_height) {
+void blit_core_naive_(uint32_t *dst, int dst_width, int dst_height, int x, int y, 
+                      uint32_t *src, int src_width, int src_height, int x0, int y0, 
+                      int w, int h) {
   
   for (int yoff = 0; yoff < h; yoff++) {
     int y1 = src_height - y0 - yoff;
@@ -112,13 +129,17 @@ void blit_core_naive_(uint32_t *dst, int x, int y, int dst_width, int dst_height
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Blit sprites into raster [R interface]
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP blit_(SEXP nr_, SEXP x_, SEXP y_, SEXP src_, SEXP x0_, SEXP y0_, SEXP w_, SEXP h_,
-           SEXP hjust_, SEXP vjust_, SEXP respect_alpha_) {
+SEXP nr_blit_(SEXP dst_  , SEXP src_, 
+             SEXP x_    , SEXP y_, 
+             SEXP x0_   , SEXP y0_, 
+             SEXP w_    , SEXP h_,
+             SEXP hjust_, SEXP vjust_, 
+             SEXP respect_alpha_) {
   
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   // Sanity checks
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  assert_nativeraster(nr_);
+  assert_nativeraster(dst_);
   assert_nativeraster(src_);
   
   if (!(Rf_length(x_) == 1 || Rf_length(y_) == 1 || Rf_length(x_) == Rf_length(y_))) {
@@ -126,10 +147,12 @@ SEXP blit_(SEXP nr_, SEXP x_, SEXP y_, SEXP src_, SEXP x0_, SEXP y0_, SEXP w_, S
   }
   
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  // Destination dimensions
+  // Dimensions
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  int nr_width  = Rf_ncols(nr_);
-  int nr_height = Rf_nrows(nr_);
+  int dst_width  = Rf_ncols(dst_);
+  int dst_height = Rf_nrows(dst_);
+  int src_width  = Rf_ncols(src_);
+  int src_height = Rf_nrows(src_);
   
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   // Ensure the coordinates are integers
@@ -145,8 +168,6 @@ SEXP blit_(SEXP nr_, SEXP x_, SEXP y_, SEXP src_, SEXP x0_, SEXP y0_, SEXP w_, S
   int x0 = Rf_asInteger(x0_);
   int y0 = Rf_asInteger(y0_);
   
-  int src_width  = Rf_ncols(src_);
-  int src_height = Rf_nrows(src_);
   
   int w = Rf_asInteger(w_) < 0 ? src_width  : Rf_asInteger(w_);
   int h = Rf_asInteger(h_) < 0 ? src_height : Rf_asInteger(h_);
@@ -156,17 +177,20 @@ SEXP blit_(SEXP nr_, SEXP x_, SEXP y_, SEXP src_, SEXP x0_, SEXP y0_, SEXP w_, S
           x0, y0, w, h, src_width, src_height);
   }
   
-  int hjust = (int)round(Rf_asReal(hjust_) * src_width);
-  int vjust = (int)round(Rf_asReal(vjust_) * src_height);
+  int hjust = (int)round(Rf_asReal(hjust_) * (src_width  - 1));
+  int vjust = (int)round(Rf_asReal(vjust_) * (src_height - 1));
   
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   // Blit mulitple copies
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  uint32_t *nr  = (uint32_t *)INTEGER(nr_);
+  uint32_t *dst = (uint32_t *)INTEGER(dst_);
   uint32_t *src = (uint32_t *)INTEGER(src_);
   bool respect_alpha = Rf_asLogical(respect_alpha_);
   for (int i = 0; i < Rf_length(x_); i++) {
-    blit_core_(nr, x[i] - hjust, y[i] - vjust, nr_width, nr_height, src, x0, y0, w, h, src_width, src_height, respect_alpha);
+    blit_core_(
+      dst, dst_width, dst_height, x[i] - hjust, y[i] - vjust, 
+      src, src_width, src_height,           x0,           y0, 
+      w, h, respect_alpha);
   }
   
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -174,86 +198,118 @@ SEXP blit_(SEXP nr_, SEXP x_, SEXP y_, SEXP src_, SEXP x0_, SEXP y0_, SEXP w_, S
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   if (freex) free(x);
   if (freey) free(y);
-  return nr_;
+  return dst_;
 }
 
 
-SEXP blit_list_(SEXP nr_, SEXP x_, SEXP y_, SEXP src_list_, SEXP src_idx_, SEXP hjust_, SEXP vjust_, SEXP respect_alpha_) {
-  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  // Sanity checks
-  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  assert_nativeraster(nr_);
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// 
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+SEXP nr_blit_bulk_(SEXP dst_, SEXP src_, SEXP config_) {
   
-  if (!Rf_isNewList(src_list_)) {
-    Rf_error("nr_blit_list_(): src_list_ must be a list");
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Check arguments are of correct type
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  assert_nativeraster(dst_);
+  if (!Rf_isNewList(src_)) {
+    Rf_error("'src' MUST be a list of native rasters");
+  }
+  int Nsrc = Rf_length(src_);
+  for (int i = 0; i < Nsrc; ++i) {
+    assert_nativeraster(VECTOR_ELT(src_, i));
+  }
+  if (!Rf_inherits(config_, "data.frame")) {
+    Rf_error("Config must be a data.frame");
   }
   
-  if (!(Rf_length(x_) == 1 || Rf_length(y_) == 1 || Rf_length(x_) == Rf_length(y_))) {
-    Rf_error("'x' and 'y' must be the length 1 or N.  x = %i, y = %i", Rf_length(x_), Rf_length(y_));
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Unpack the config arguments and convert to correct type if necessary
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  int N = Rf_length(VECTOR_ELT(config_, 0));
+  if (N == 0) {
+    Rf_warning("Empty 'config'");
+    return dst_;
   }
   
+  SEXP idx_           = get_df_col_or_error(config_, "idx");
+  SEXP x_             = get_df_col_or_error(config_, "x");
+  SEXP y_             = get_df_col_or_error(config_, "y");
+  SEXP w_             = get_df_col_or_error(config_, "w");
+  SEXP h_             = get_df_col_or_error(config_, "h");
+  SEXP x0_            = get_df_col_or_error(config_, "x0");
+  SEXP y0_            = get_df_col_or_error(config_, "y0");
+  SEXP hjust_         = get_df_col_or_error(config_, "hjust");
+  SEXP vjust_         = get_df_col_or_error(config_, "vjust");
+  SEXP respect_alpha_ = get_df_col_or_error(config_, "respect_alpha");
+  SEXP render_        = get_df_col_or_error(config_, "render");
+  
+  bool freeidx = false;
+  bool freex = false, freey = false, freex0 = false, freey0 = false;
+  bool freew = false, freeh = false;
+  
+  int *idx   = as_int32_vec(idx_, N, &freeidx);
+  int *x     = as_int32_vec(x_  , N, &freex);
+  int *y     = as_int32_vec(y_  , N, &freey);
+  int *w     = as_int32_vec(w_  , N, &freew);
+  int *h     = as_int32_vec(h_  , N, &freeh);
+  int *x0    = as_int32_vec(x0_ , N, &freex0);
+  int *y0    = as_int32_vec(y0_ , N, &freey0);
+  
+  double *hjust = REAL(hjust_);
+  double *vjust = REAL(vjust_);
+  int *respect_alpha = LOGICAL(respect_alpha_);
+  int *render        = LOGICAL(render_);
   
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  // Destination dimensions
+  // Do the loop
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  int nr_width  = Rf_ncols(nr_);
-  int nr_height = Rf_nrows(nr_);
+  uint32_t *dst = (uint32_t *)INTEGER(dst_);
+  int dst_width  = Rf_ncols(dst_);
+  int dst_height = Rf_nrows(dst_);
   
-  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  // Ensure the coordinates are integers
-  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  bool freex = false, freey = false;
-  int N = calc_max_length(2, x_, y_);
-  int *x = as_int32_vec(x_, N, &freex);
-  int *y = as_int32_vec(y_, N, &freey);
-  
-  bool freeindices = false, free_respect_alpha = false;
-  int *indices = as_int32_vec(src_idx_, N, &freeindices);
-  int *respect_alpha = as_int32_vec(respect_alpha_, N, &free_respect_alpha);
-  
-  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  // List of native rasters
-  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  bool freehjust = false, freevjust = false;
-  double *hjust = as_double_vec(hjust_, N, &freehjust);
-  double *vjust = as_double_vec(vjust_, N, &freevjust);
-  
-  uint32_t *nr  = (uint32_t *)INTEGER(nr_);
-  
-  for (int i = 0; i < N; i++) {
-    int idx = indices[i] - 1;
-    if (idx >= Rf_length(src_list_) || idx < 0) {
-      Rprintf("nr_blit_list_(): 'src_idx' is out of bounds for given 'src_list'. %i is not in range [0, %i]\n", idx, Rf_length(src_list_) - 1);
-      continue;
+  for (int i = 0; i < N; ++i) {
+    if (!render[i]) continue;
+    int this_idx = idx[i] - 1;
+    if (idx[i] < 0 || idx[i] >= Nsrc) {
+      Rf_error("idx = %i at row %i is invalid. Not in range [1, %i]", idx[i], i, N);
+      // continue;
     }
-    SEXP src_ = VECTOR_ELT(src_list_, idx);
-    assert_nativeraster(src_);
-    uint32_t *src = (uint32_t *)INTEGER(src_);
+    SEXP this_src_ = VECTOR_ELT(src_, this_idx);
+    uint32_t *src = (uint32_t *)INTEGER(this_src_);
+    int src_width  = Rf_ncols(this_src_);
+    int src_height = Rf_nrows(this_src_);
 
-    int src_width  = Rf_ncols(src_);
-    int src_height = Rf_nrows(src_);
+    int this_hjust = (int)round(hjust[i] * (src_width  - 1));
+    int this_vjust = (int)round(vjust[i] * (src_height - 1));
 
-    int this_hjust = (int)round(hjust[i] * src_width);
-    int this_vjust = (int)round(vjust[i] * src_height);
-
-    blit_core_(nr, x[i] - this_hjust, y[i] - this_vjust, nr_width, nr_height,
-               src, 0, 0, src_width, src_height, src_width, src_height,
-               (bool)respect_alpha[i]);
+    int this_w = w[i] < 0 ? src_width  : w[i];
+    int this_h = h[i] < 0 ? src_height : h[i];
     
-    
+    blit_core_(
+      dst, dst_width, dst_height, x[i] - this_hjust, y[i] - this_vjust,
+      src, src_width, src_height,           x0[i],           y0[i],
+      this_w, this_h, respect_alpha[i]);
   }
+  
   
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   // Tidy and return
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  if (freeindices) free(indices);
-  if (free_respect_alpha) free(respect_alpha);   
-  if (freehjust) free(hjust); 
-  if (freevjust) free(vjust);   
-  if (freex) free(x);
-  if (freey) free(y);
-  return nr_;
+  if (freeidx) free(idx);
+  if (freex)   free(x);
+  if (freey)   free(y);
+  if (freew)   free(w);
+  if (freeh)   free(h);
+  if (freex0)  free(x0);
+  if (freey0)  free(y0);
+  return dst_;
 }
+
+
+
+
+